Moved the test to api/key_entry/.
